Move unit-ordering sort functions to a public location
Move sort_by_unit_name, sort_by_display_name, sort_by_dimensionality,
and the SortFunc type alias out of the private
pint.delegates.formatter._compound_unit_helpers module into a new
pint.delegates.formatter.sorting module, re-exported from
pint.delegates.formatter. This makes them discoverable in the API
reference (docs/api/facets.rst already automodules that package) and
importable without reaching into a private module.

Also documents the required default_sort_func signature in
docs/user/formatting.rst and updates its example to import from the
new public location.
