+++ CHANGES
@@ -10,6 +10,7 @@ Pint Changelog
 - Raise ``DefinitionSyntaxError`` instead of a bare ``AssertionError`` when parsing an expression that is only an operator, e.g. ``ureg('-')`` or ``ureg('*')`` (#2124)
 - Document registry preprocessors with an example for parsing non-standard square-unit spellings (#1211, #2062)
 - Document how to control the order of units within a compound unit's string representation via `default_sort_func` and `dim_order` (#2086)
+- Move `sort_by_unit_name`, `sort_by_display_name`, `sort_by_dimensionality` and the `SortFunc` type alias from the private `pint.delegates.formatter._compound_unit_helpers` module to the public `pint.delegates.formatter` module (#2086)
 - Fix format specs with a width/fill/alignment (e.g. `f"{q:20.1f}"`) not honoring the requested field width for quantities, because the width was applied only to the magnitude before the unit was appended. The width/fill/alignment now applies to the whole formatted quantity while sign/precision/type/zero-padding stay with the magnitude (#1798)
 - Support the superscript decimal point `⋅` (U+22C5) and fraction slash `⸍` (U+2E0D) in unit exponents, e.g. `gr⁰⋅³³³` (`grain**0.333`) and `gr¹⸍³` (`grain**(1/3)`) (#2250)
 - Fix `KeyError('')` raised when the `dimensionless` unit is referenced by name, e.g. `get_dimensionality("dimensionless")`, `Quantity(5, "dimensionless").check("dimensionless")`, or a unit defined via `* dimensionless` (#1513, #1994)
